Get new q and k values
By default, k = 3.5 and p = k -3 + q. Since q is very sharp for a bimodal distribution (i.e., > 0.5) having k = 3.5 would result in a p value that is >= 1. The problem is that the equations for surface and volume density distribution has a factor of (1 - p), resulting in a distribution that is >= 0 (i.e., not physical). This update helps calculate new p and k values such that p <= 0.5